// src/supernode/enums/sn-bonus-type.enum.ts
export enum SN_BONUS_TYPE {
  BASE_REFERRAL = 'base-referral',
  BUILDER_GENERATIONAl = 'builder-generational',
  BUILDER_REFERRAL = 'builder-referral',
  SUPERNODE_GLOBAL_POOL = 'supernode-global-pool',
  MATCHING_BONUS = 'matching-bonus',
}

export enum SN_BONUS_SUMMARY_TYPE {
  BASE_REFERRAL = 'base-referral',
  BUILDER_GENERATIONAl = 'builder-generational',
  BUILDER_REFERRAL = 'builder-referral',
  MATCHING_BONUS = 'matching-bonus',
  TOTAL_REWARDS = 'total-reward',
}

export enum SN_BONUS_SUMMARY_TYPE_REPORT {
  BASE_REFERRAL = 'base-referral',
  BUILDER_GENERATIONAl = 'builder-generational',
  BUILDER_REFERRAL = 'builder-referral',
  MATCHING_BONUS = 'matching-bonus',
}